Move BreakTag into .doc, make HasUniqueId package-private
