+++ palantir-java-format/src/main/java/com/palantir/javaformat/doc/BreakTag.java
@@ -1,0 +1,7 @@
+package com.palantir.javaformat.doc;
+
+import com.google.errorprone.annotations.Immutable;
+
+/** Unique identifier for a break. */
+@Immutable
+public final class BreakTag extends HasUniqueId {}
