+++ palantir-java-format/src/main/java/com/palantir/javaformat/doc/HasUniqueId.java
@@ -3,12 +3,12 @@ package com.palantir.javaformat.doc;
 import fj.Ord;
 import java.util.concurrent.atomic.AtomicInteger;
 
-public abstract class HasUniqueId {
+abstract class HasUniqueId {
     private static final AtomicInteger UNIQUE_ID_GENERATOR = new AtomicInteger();
 
-    public final int uniqueId = UNIQUE_ID_GENERATOR.incrementAndGet();
+    final int uniqueId = UNIQUE_ID_GENERATOR.incrementAndGet();
 
-    public static <D extends HasUniqueId> Ord<D> ord() {
+    static <D extends HasUniqueId> Ord<D> ord() {
         return Ord.on((D obj) -> obj.uniqueId, Ord.intOrd).ord();
     }
 }
