+++ palantir-java-format/src/main/java/com/palantir/javaformat/java/JavaInputAstVisitor.java
@@ -67,8 +67,8 @@ import com.palantir.javaformat.Op;
 import com.palantir.javaformat.OpenOp;
 import com.palantir.javaformat.OpsBuilder;
 import com.palantir.javaformat.OpsBuilder.BlankLineWanted;
-import com.palantir.javaformat.Output.BreakTag;
 import com.palantir.javaformat.doc.Break;
+import com.palantir.javaformat.doc.BreakTag;
 import com.palantir.javaformat.doc.FillMode;
 import com.palantir.javaformat.doc.Token;
 import com.palantir.javaformat.java.DimensionHelpers.SortedDims;
